+++ Source/EventFlow.Tests/UnitTests/ReadStores/AsyncReadModelPopulatorTests.cs
@@ -1,0 +1,6 @@
+﻿namespace EventFlow.Tests.UnitTests.ReadStores
+{
+    public class AsyncReadModelPopulatorTests : BaseReadModelTests<TestAsyncReadModel>
+    {
+    }
+}
