+++ Source/EventFlow.Tests/UnitTests/ReadStores/SingleAggregateReadStoreManagerTests.cs
@@ -29,8 +29,8 @@ namespace EventFlow.Tests.UnitTests.ReadStores
 {
     [Category(Categories.Unit)]
     public class SingleAggregateReadStoreManagerTests : ReadStoreManagerTestSuite<SingleAggregateReadStoreManager<
-        IReadModelStore<TReadModel>,
-        TReadModel>>
+        IReadModelStore<TestReadModel>,
+        TestReadModel>>
     {
     }
 }
